Move checking business logic to engine
CMK-12298
